Extract chokidar extraction into its own file
